agent-integration/python-agents/requirements.txt
openai-agents>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
asyncio>=3.4.3